Attempting to add CuArrays 1.7.0
Note, several errors appeared while testing, the most prominent of which
can be found in the RootSolvers.jl file. Here, the Verbose solution results
attempt to return something that is not a bits type (an array of indeterminite
size). Changing this to an SVector could help, but I cannot seem to get it to
work.