Create JuliaFormatter GitHub Action
On each PR, this will run the JuliaFormatter.jl on all modified .jl files, failing if the formatting modifies any of the files.

I originally tried using https://github.com/julia-actions/julia-format, but it wasn't really sufficient for our needs, so it made more sense to use a custom workflow.